firebase added

+++ deliveryApp/deliveryApp/Models/CartModel.swift
@@ -17,7 +17,7 @@ struct CartItem: Identifiable, Equatable {
     var id: String = UUID().uuidString
     
     let food: Food
-    let quantity: Int
+    let quantity: Int64
 }
 
 let cartExample: Cart = Cart(items: [CartItem(food: foodExamples[0], quantity: 1), CartItem(food: foodExamples[1], quantity: 3)])
